Пересмотрена организация хранения данных, перенос идентификаторов обратно к int, на первое время

+++ TestingApplication.Data/Repositories/DB/DbGenericRepository.cs
@@ -36,7 +36,7 @@ namespace TestingApplication.Data.Repositories.DB
             }
         }
 
-        public virtual TEntity? Get(long id)
+        public virtual TEntity? Get(int id)
         {
             return _dbSet.Find(id);
         }
